[Decomp][Player] Cohesion split: extract action-bar button methods to PlayerActionButton.cpp
